fix(api): move regular code out of test folders
Ref. metriport/metriport-internal#187
