+++ packages/api/src/command/medical/facility/delete-facility.ts
