+++ packages/api/src/command/medical/organization/delete-organization.ts
