+++ packages/api/src/external/fhir/organization/delete-organization.ts
